Round X/Y Co-Ords in infoParams. (Browser Fix)
When querying a layer in Internet Explorer - the WMS will throw an "X and Y incorrectly specified" error in the parse response. This error seems only occur in IE specifically. Investigating the X/Y parameters passed - they appear to be decimal values where integers are expected.

The fix for this is to Math.round() the X and Y values in infoParams so that the decimal value is removed by rounding the value to the closest real number before it is passed as a request when querying.